+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.1] - 2025-11-26
+### Changed
+- CreateAuditRecordCommandBuilderFactory now can also be used without the jeap-security-starter on the classpath.
+- CreateAuditRecordCommandBuilderFactory now accepts Messages int the general Message interface instead of just AvroDomainEvent.
+
 ## [2.2.0] - 2025-11-14
 ### Changed
 - update jeap-messaging from 10.0.1 to 10.1.0